% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_helpers.R
\name{chop_off_top_bottom_loading_ntiles_by_factor}
\alias{chop_off_top_bottom_loading_ntiles_by_factor}
\title{Chop off top and bottom ntile factor loadings by factor}
\usage{
chop_off_top_bottom_loading_ntiles_by_factor(factor_df, ntiles)
}
\arguments{
\item{ntiles}{how many ntiles to split the factor loadings into}

\item{df}{a data.frame in long format with at least three columns: "item_id", "factor" and "loading"}
}
\description{
Chop off top and bottom ntile factor loadings by factor
}
