% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_helpers.R
\name{chop_off_top_bottom_ntiles}
\alias{chop_off_top_bottom_ntiles}
\title{Chop off data.frame top and bottom ntile}
\usage{
chop_off_top_bottom_ntiles(df, variable, ntiles)
}
\arguments{
\item{df}{a data.frame}

\item{variable}{unquoted name of the variable to divide into ntiles by}

\item{ntiles}{how many ntiles to split variable into}
}
\description{
Takes a data.frame, splits it into ntiles and then returns only the top and bottom ntiles
}
