% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bemp_data.R
\name{count_lines}
\alias{count_lines}
\title{Quickly counts the number of lines in a list of files}
\usage{
count_lines(files, exclude_total = TRUE)
}
\arguments{
\item{files}{a vector listing the files whose line length is desired}

\item{exclude_total}{when TRUE the combined total line count across all the files is excluded
(by default wc returns the length of each file AND the sum of these lengths)}
}
\value{
A vector with the length of each file (and the sum of these lengths if exclude_total is FALSE)
}
\description{
Uses the unix wc function to count the number of lines in a list of files.
}
