% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pfmul_data.R
\name{factor_label_to_id}
\alias{factor_label_to_id}
\title{strip a robust factor label to its id}
\usage{
factor_label_to_id(factor_label)
}
\arguments{
\item{factor_label}{a character vector of robust factor labels, e.g. "factor_05"}
}
\description{
strip a robust factor label to its id
}
