% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bemp_data.R
\name{factor_sequence}
\alias{factor_sequence}
\title{create a sequence of factor labels}
\usage{
factor_sequence(length_out, sequence_type = "robust")
}
\arguments{
\item{length_out}{length of the sequence}

\item{sequence_type}{either "robust" (lowercase, underscored and zero-padded) or
"pretty" (to be used in plotting)}
}
\description{
create a sequence of factor labels
}
