% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pfmul_normalizations.R
\name{frequency_normalize_pfmul_loadings}
\alias{frequency_normalize_pfmul_loadings}
\title{Normalize PFMUL factor loadings within user/item}
\usage{
frequency_normalize_pfmul_loadings(factor_df)
}
\arguments{
\item{factor_df}{a data.frame in long format with at least three columns: "user_id" or "item_id", "factor_id" and "loading"}
}
\description{
Normalize PFMUL factor loadings within user/item
}
