% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bemp_data.R
\name{get_bemp_checkpoints}
\alias{get_bemp_checkpoints}
\title{Returns a list of the available checkpoints for a BEMP run}
\usage{
get_bemp_checkpoints(model_path)
}
\arguments{
\item{model_path}{the directory in which the results of the BEMP model run reside}
}
\value{
a data.table with all existing checkpoints (i.e. innerProduct files)
for the selected run, with information from the validation set scores merged on
as well as the line count of the innerProduct files, since this can help detect incomplete checkpoints
which can occur if a BEMP run is interrupted or crashes.
}
\description{
Returns a list of the available checkpoints for a BEMP run
}
