% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bemp_data.R
\name{get_bemp_inner_products}
\alias{get_bemp_inner_products}
\title{Get pre-computed inner products for the BEMP model}
\usage{
get_bemp_inner_products(model_path, iteration)
}
\arguments{
\item{model_path}{the directory in which the results of the BEMP model run reside}

\item{iteration}{the iteration}
}
\value{
for each `user_id` - `item_id` combination
\itemize{
  \item alpha1 = lambda0_i + theta_u * alpha_i + obsItem_u * obsItem_i
  \item alpha2 = lambda0_i + theta_u * alpha_i + obsItem_u * obsItem_i + mu_i * delta_w (the average delta_w across all w)
  \item eta = gamma_u * beta_i
}
}
\description{
Get pre-computed inner products for the BEMP model
}
