% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bemp_data.R
\name{get_bemp_performance_measures}
\alias{get_bemp_performance_measures}
\title{Get BEMP performance measures}
\usage{
get_bemp_performance_measures(data_dir)
}
\arguments{
\item{data_dir}{the directory in which the results of the BEMP model run reside}
}
\value{
a tibble with the following columns:
\itemize{
  \item iteration
  \item duration_seconds
  \item log_likelihood
  \item accuracy: the fraction of correctly classified instances.
  \item \eqn{precision_i}: the fraction of instances where we correctly declared \eqn{i} out of all instances where the algorithm declared \eqn{i} (then I take the average across all \eqn{i})
  \item recall: the fraction of instances where we correctly declared \eqn{i} out of all of the cases where the true of choice was \eqn{i} (also averaged across all \eqn{i})
  \item F1-score: defined according to the formula \eqn{f1_score = 2 * precision * recall / (precision + recall)}
  \item total_instances
}
}
\description{
Get BEMP performance measures
}
