% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pfmul_data.R
\name{get_distance_coefficients}
\alias{get_distance_coefficients}
\title{Extracts the distance coefficients from the PFMUL output}
\usage{
get_distance_coefficients(data_dir, shape = "long")
}
\arguments{
\item{data_dir}{the directory in which the parameter tsv files reside}

\item{shape}{"matrix" if the raw user x item coefficient matrix should be returned, "long" if the coefficients are to be returned as a tidy (long) data.frame}
}
\description{
Extracts the distance coefficients from the PFMUL output
}
