% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bemp_data.R
\name{get_icvar_names}
\alias{get_icvar_names}
\title{Get names of the BEMP ICvars}
\usage{
get_icvar_names(model_path)
}
\arguments{
\item{model_path}{the directory in which the results of the BEMP model run reside}
}
\description{
Get names of the BEMP ICvars
}
