% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_helpers.R
\name{get_id_col}
\alias{get_id_col}
\title{Get the data.frame ID column}
\usage{
get_id_col(df)
}
\arguments{
\item{df}{a data.frame with either user_id or item_id}
}
\description{
Get the data.frame ID column
}
