% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/benchmark_data.R
\name{get_model_predictions}
\alias{get_model_predictions}
\title{Gets predictions from bemp, stata, or other models}
\usage{
get_model_predictions(model_type, model_path, input_data_path,
  selected_iteration = 0, verbose = FALSE)
}
\arguments{
\item{model_type}{Can be bemp, stata, or anything else (as long as it has a predictions.tsv file)}

\item{model_path}{The folder containing the outputs from the model}

\item{input_data_path}{The folder that contains the bemp-style model inputs}

\item{selected_iteration}{For bemp models only. Which checkpoint to use.}

\item{verbose}{}
}
\value{
A data.table with the predictions from the model
}
\description{
Gets predictions from bemp, stata, or other models
}
