% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bemp_data.R
\name{get_sessions}
\alias{get_sessions}
\title{Get the session data}
\usage{
get_sessions(data_dir, samples = c("train", "test", "validation"),
  verbose = FALSE)
}
\arguments{
\item{verbose}{print messages along the way}

\item{model_path}{the directory in which the raw data reside}

\item{sample}{"train", "validation", "test", any combination thereof, or "all" as a shorthand}
}
\description{
Get the session data
}
