% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/benchmark_data.R
\name{get_stata_model_internals}
\alias{get_stata_model_internals}
\title{Get the Stata model internals}
\usage{
get_stata_model_internals(model_path, samples = c("train", "test",
  "validation"), cols = c("sample", "session_id", "user_id", "item_id",
  "utility"), verbose = FALSE)
}
\arguments{
\item{model_path}{the directory in which the results of the Stata model run reside}

\item{cols}{the columns to return. Choose only those that are necessary for a speedup}

\item{verbose}{print messages along the way}

\item{sample}{"train", "validation", "test", any combination thereof, or "all" as a shorthand}
}
\value{
data.table containing
\itemize{
  \item sample
  \item session_id
  \item user_id
  \item item_id
  \item chosen whether the item was chosen
  \item utility utility unter the model
  \item choice_prob choice probability under the model
}
}
\description{
Get the Stata model internals
}
