% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_helpers.R
\name{get_top_bottom_ntile_by_factor}
\alias{get_top_bottom_ntile_by_factor}
\title{Get top and bottom ntile items}
\usage{
get_top_bottom_ntile_by_factor(factor_df, ntiles)
}
\arguments{
\item{factor_df}{a data.frame in long format with at least three columns: "item_id", "factor" and "loading"}

\item{ntiles}{how many ntiles to split the factor loadings into}
}
\description{
Get top and bottom ntile items
}
