% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pfmul_data.R
\name{get_utility_components}
\alias{get_utility_components}
\title{Extract utility components from the PFMUL output}
\usage{
get_utility_components(component, data_dir, iteration = NULL,
  shape = "long")
}
\arguments{
\item{component}{character: either "latent_factors" or "distance"}

\item{data_dir}{the directory in which the parameter tsv files reside}

\item{iteration}{integer: the iteration at which to evaluate the parameters}

\item{shape}{"matrix" if the raw user x item coefficient matrix should be returned, "long" if the coefficients are to be returned as a tidy (long) data.frame}
}
\description{
Extract utility components from the PFMUL output
}
