% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logits.R
\name{logit_heatmap}
\alias{logit_heatmap}
\title{logit heatmap}
\usage{
logit_heatmap(factor_df, covariate_df, covariates, ntiles = 10)
}
\arguments{
\item{factor_df}{a data.frame in long format with at least three columns: "item_id", "factor" and "loading"}

\item{covariate_df}{a data.frame with covariates that will join against factor_df on item_id}

\item{covariates}{character vector of covariate names to be used as independent variables in the logit}

\item{ntiles}{how many ntiles to split the factor loadings into}
}
\description{
Tests for differences in mean factor loadings between top and bottom ntile by running logit regressions on a set of covariates
}
