% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logits.R
\name{logit_predict_top_ntile}
\alias{logit_predict_top_ntile}
\title{Predict membership in the top ntile}
\usage{
logit_predict_top_ntile(df, covariates)
}
\arguments{
\item{df}{a data.frame that must contain both a "top_ntile" variable and all covariates}

\item{covariates}{character vector of variable names}
}
\description{
Predict membership in the top ntile given a set of covariates
}
