% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize.R
\name{normalize_factor_loadings}
\alias{normalize_factor_loadings}
\title{Normalize factor loadings}
\usage{
normalize_factor_loadings(factor_df, method)
}
\arguments{
\item{factor_df}{a data.frame in long format with at least three columns: "user_id" or "item_id", "factor" and "loading"}

\item{method}{the normalization method. Possible options:
- raw
- demean
- exp_normalized
- percent}
}
\description{
Normalize factor loadings using different methods
}
