% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bemp_data.R
\name{parse_bemp_label}
\alias{parse_bemp_label}
\title{Parse the options the bemp model was estimated with given a directory name}
\usage{
parse_bemp_label(description)
}
\arguments{
\item{label}{the label of the bemp model, often the directory name under which it is saved}
}
\value{
tibble with parameter:value pairs
}
\description{
Parse the options the bemp model was estimated with given a directory name
}
