% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/t_tests.R
\name{perform_t_tests}
\alias{perform_t_tests}
\title{t-test for differences in covariates}
\usage{
perform_t_tests(factor_df, covariate_df, covariates, ntiles)
}
\arguments{
\item{factor_df}{a data.frame in long format with at least three columns: "item_id", "factor" and "loading"}

\item{covariate_df}{a data.frame with covariates that will join against factor_df on item_id}

\item{covariates}{character vector of covariate names to perform the t-tests within}

\item{ntiles}{how many ntiles to split the factor loadings into}
}
\description{
Performs the t-tests for t_test_heatmap
}
