% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bemp_data.R
\name{read_bemp_parameter_file}
\alias{read_bemp_parameter_file}
\title{Read in bemp parameter file}
\usage{
read_bemp_parameter_file(parameter_name, moment = "mean", data_dir,
  iteration = NULL, shape = "long")
}
\arguments{
\item{parameter_name}{the (character) name of the parameter to be read in, e.g. "alpha" for item factor loadings}

\item{moment}{the moment of the parameter to be read in, either "mean" or "std"}

\item{data_dir}{the directory in which the results of the BEMP model run reside}

\item{iteration}{integer: the iteration at which to evaluate the parameters}

\item{shape}{character: whether to return the parameters in 'long' or 'wide' format}
}
\description{
Read in bemp parameter file
}
