% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pfmul_data.R
\name{read_pfmul_parameter_file}
\alias{read_pfmul_parameter_file}
\title{Read in PFMUL parameter file}
\usage{
read_pfmul_parameter_file(parameter_name, moment = "mean", data_dir,
  shape = "long")
}
\arguments{
\item{parameter_name}{the (character) name of the parameter to be read in, e.g. "alpha" for item factor loadings}

\item{moment}{the moment of the parameter to be read in, either "mean" or "std"}

\item{data_dir}{the directory in which the parameter tsv files reside}

\item{shape}{whether to return the parameters in long or wide format}
}
\description{
Read in PFMUL parameter file
}
