% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/benchmark_data.R
\name{rename_if_possible}
\alias{rename_if_possible}
\title{If there is a column named old, rename it new}
\usage{
rename_if_possible(dt, old, new)
}
\description{
If there is a column named old, rename it new
}
