% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{replace_name}
\alias{replace_name}
\title{Replace name}
\usage{
replace_name(dt, old_name, new_name)
}
\arguments{
\item{dt}{a data.table}

\item{old_name}{if dt has a column with this name, rename it new_name, otherwise do nothing}

\item{new_name}{the new name to replace old_name}
}
\description{
If dt has column old_name, then rename that column new_name by reference
}
