% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logits.R
\name{run_logits}
\alias{run_logits}
\title{Logits for differences in covariates}
\usage{
run_logits(factor_df, covariate_df, covariates, ntiles)
}
\arguments{
\item{factor_df}{a data.frame in long format with at least three columns: "item_id" or "user_id", "factor" and "loading"}

\item{covariate_df}{a data.frame with covariates that will join against factor_df on the ID column}

\item{covariates}{character vector of covariate names to perform the t-tests within}

\item{ntiles}{how many ntiles to split the factor loadings into}
}
\description{
Performs the logit estimations for logit_heatmap
}
