% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/top_n.R
\name{top_n_table}
\alias{top_n_table}
\title{Top n by factor}
\usage{
top_n_table(factor_df, covariate_df, include_columns = NULL, n = 10)
}
\arguments{
\item{factor_df}{a data.frame in long format with at least three columns: "item_id", "factor" and "loading"}

\item{covariate_df}{a data.frame with covariates that will join against factor_df on item_id}

\item{include_columns}{character vector of variable names to include in the table}

\item{n}{how many items to return per factor}
}
\description{
Returns the top n items/users by loading
}
