% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bemp_data.R
\name{get_bemp_model_internals}
\alias{get_bemp_model_internals}
\title{Get the BEMP model internals}
\usage{
get_bemp_model_internals(model_path, iteration)
}
\arguments{
\item{model_path}{the directory in which the results of the BEMP model run reside}

\item{iteration}{the iteration}
}
\value{
data.table containing
\itemize{
  \item session_id
  \item user_id
  \item item_id
  \item alpha1 = lambda0_i + theta_u * alpha_i + obsItem_u * obsItem_i
  \item alpha2 = lambda0_i + theta_u * alpha_i + obsItem_u * obsItem_i + mu_i * delta_w (the average delta_w across all w)
  \item eta = gamma_u * beta_i
  \item distance distance between item and user in km
  \item chosen whether the item was chosen
  \item utility utility unter the model: alpha2 - eta * log(distance)
  \item choice_prob choice probability under the model
}
}
\description{
Get the BEMP model internals
}
