% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bemp_data.R
\name{get_sessions}
\alias{get_sessions}
\title{Get the session data}
\usage{
get_sessions(model_path, iteration, samples = c("train", "test",
  "validation"), verbose = FALSE)
}
\arguments{
\item{model_path}{the directory in which the results of the BEMP model run reside}

\item{iteration}{the iteration}

\item{verbose}{print messages along the way}

\item{sample}{"train", "validation", "test", any combination thereof, or "all" as a shorthand}
}
\description{
Get the session data
}
