% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bemp_data.R
\name{parse_bemp_logfile}
\alias{parse_bemp_logfile}
\title{Parse the BEMP logfile}
\usage{
parse_bemp_logfile(model_path)
}
\arguments{
\item{model_path}{the directory in which the results of the BEMP model run reside}
}
\value{
nested list
}
\description{
Parse the BEMP logfile
}
